requirements.txt
Keras==2.1.6
Markdown==2.6.11
Pillow==5.1.0
PyYAML==3.12
Theano==1.0.1
Werkzeug==0.14.1
absl-py==0.2.1
astor==0.6.2
bleach==1.5.0
cycler==0.10.0
gast==0.2.0
grpcio==1.12.0
h5py==2.7.1
html5lib==0.9999999
kiwisolver==1.0.1
matplotlib==2.2.2
numpy==1.14.0
opencv-python==3.4.0.12
pip==9.0.1
protobuf==3.5.2.post1
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
scipy==1.1.0
setuptools==38.5.1
six==1.11.0
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
wheel==0.30.0